Move the tests over to the index file
We should not create a nomocks file because we can just group the tests under a unique describe block of test/index.ts.
